+++ jaguar_serializer_cli/test/serializer/fromMap/dynamic.dart
@@ -15,6 +15,20 @@ main() {
               new DynamicModel(
                   scalar: 5, list: ['5', 'five'], map: {'5': 'five'}));
         });
+        test("Pass", () {
+          var dt1 = new DateTime.now();
+          var dt2 = dt1.add(new Duration(minutes: 1));
+          expect(
+              new DynamicPassSerializer().fromMap({
+                'scalar': dt1,
+                'list': ['5', 'five', dt1],
+                'map': {'5': 'five', 'date': dt2}
+              }),
+              new DynamicModel(
+                  scalar: dt1,
+                  list: ['5', 'five', dt1],
+                  map: {'5': 'five', 'date': dt2}));
+        });
       });
     });
   });