+++ jaguar_serializer_cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.16
+
++ dynamic field
+
+## 2.1.15
+
++ Bug fix: [Final non-nullable contructor](https://github.com/Jaguar-dart/jaguar_serializer/issues/144)
+
 ## 2.1.14
 
 + Fixed bug when source is `Map<dynamic, dynamic>`